Optimization

+++ src/ru/erked/sflight/tech/SFButtonA.java
@@ -20,8 +20,8 @@ public class SFButtonA {
 	private Camera camera;
 	
 	public SFButtonA(String texture, float width, float x, float y, Camera camera){
-		this.x = x - Gdx.graphics.getWidth()/2;
-		this.y = y - Gdx.graphics.getHeight()/2;
+		this.x = x - (float)Gdx.graphics.getWidth()/2.0F;
+		this.y = y - (float)Gdx.graphics.getHeight()/2.0F;
 		textureI = new Texture(texture + "I.png");
 		textureA = new Texture(texture + "A.png");
 		aspect = (float)textureA.getWidth()/(float)textureI.getWidth();
@@ -78,7 +78,7 @@ public class SFButtonA {
 		if(bool){
 			textureCur = textureA;
 			sprite.setTexture(textureCur);
-			sprite.setBounds(x, y, width*aspect, height*aspect);
+			sprite.setBounds(x, y, (float)width*aspect, (float)height*aspect);
 		}else{
 			textureCur = textureI;
 			sprite.setTexture(textureCur);
@@ -105,8 +105,8 @@ public class SFButtonA {
 	
 	public void setCoordinates(){
 		if(isActiveMode()){
-			float prevX = camera.position.x + (x - width/aspect);
-			float prevY = camera.position.y + (y - height/aspect);
+			float prevX = camera.position.x + (x - (float)width/aspect);
+			float prevY = camera.position.y + (y - (float)height/aspect);
 			sprite.setX(prevX);
 			sprite.setY(prevY);
 		}else{
